fix: tailwind demo
