+++ packages/expo-styling/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "main": "./src/index",
   "scripts": {
-    "build": "expo-module typecheck --noEmit false",
+    "build": "expo-module typecheck",
     "clean": "expo-module clean",
     "lint": "expo-module lint",
     "test": "expo-module test",
