+++ packages/expo-styling/src/runtime/native/utils.ts
@@ -3,32 +3,12 @@ import React from "react";
 import { Easing } from "react-native-reanimated";
 
 import { exhaustiveCheck } from "../../css-to-rn/utils";
-import { RuntimeValue, SignalLike } from "../../types";
+import { SignalLike } from "../../types";
 
 export function unwrap<T>(value: T | SignalLike<T>): T {
   return value && typeof value === "object" && "get" in value
     ? value.get()
     : value;
-}
-
-export function kebabToCamelCase(str: string) {
-  return str.replace(/-./g, (x) => x[1].toUpperCase());
-}
-
-export function isRuntimeValue(value: unknown): value is RuntimeValue {
-  if (!value) {
-    return false;
-  } else if (Array.isArray(value)) {
-    return value.some((v) => isRuntimeValue(v));
-  } else if (typeof value === "object") {
-    if ("type" in value && value.type === "runtime") {
-      return true;
-    } else {
-      return Object.values(value).some((v) => isRuntimeValue(v));
-    }
-  } else {
-    return false;
-  }
 }
 
 function useDynamicDependancies(value: React.DependencyList) {
